requirements.txt
pendulum
psutil
discord.py
loguru